Don't steal focus when debugger attachment fails
Usually this fails because the engine was unable to launch the game for
whatever reason (invalid manifest, e.g.), and having the IDE steal back
focus to note the failure is annoying if the user happens to be in the
middle of reading the error text in the terminal.  Now it's more subtle
about it, showing the notice in the status bar instead of a popup
window.