Fix crash when using "Save All" with Start Page up
The IDE was trying to call StartPageView's Save() method, which failed
because it doesn't implement it. However, it's also marked as
non-saveable (CanSave = false), so not being implemented is expected in
that case. Fixed by adding a check to the default implementation of
Save() to keep quiet for non-saveable views.