Fix F5 not working after a debugging session ends
All the keyboard shortcuts for the IDE are associated with menu
commands; since part of my refactoring was to not check for menu item
enabled status until the corresponding menu is opened, this could cause
shortcut keys to get disabled in situations where they shouldn't have
been.  A rather annoying bug, but thankfully, easy to fix.